Extract mapping to its own service
